backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
apscheduler==3.10.4
qrcode[pil]==7.4.2
Pillow==10.3.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
python-dotenv==1.0.1
httpx==0.27.0
reportlab==4.2.2
firebase-admin==6.5.0
openpyxl==3.1.2
